feat: restructure header component with react responsive and removing refreshspin icon
