+++ package.json
@@ -21,6 +21,7 @@
     "react": "17.0.2",
     "react-dom": "17.0.2",
     "react-query": "3.32.1",
+    "react-responsive": "9.0.0-beta.4",
     "sharp": "0.29.2"
   },
   "devDependencies": {
