--- src/components/Layout/Header/RefreshSpin/RefreshSpin.module.css
@@ -1,6 +1,0 @@
-.refreshContainer {
-  @apply absolute rounded-full shadow bg-white p-1;
-  transform: translate(-50%, -50%);
-  left: 50%;
-  top: 100%;
-}
