+++ src/components/Layout/Header/mobile/Header.module.css
@@ -1,0 +1,39 @@
+.logoWrapper {
+  @apply relative;
+  width: 163px;
+  height: 28px;
+}
+
+.hamburgerMenuWrapper {
+  @apply relative;
+  width: 24px;
+  height: 14px;
+}
+
+.mobileSearchWrapper {
+  @apply relative;
+  width: 24px;
+  height: 24px;
+}
+
+.mobileNavigationMenu {
+  @apply inline-flex items-center justify-center p-2 rounded-lg;
+  @apply hover:text-white focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-offset-blue-800 focus:ring-white;
+  @apply md:hidden;
+}
+
+.searchWrapper {
+  @apply relative;
+  width: 20px;
+  height: 20px;
+}
+
+.search {
+  @apply flex-1 appearance-none w-full py-2 px-4 bg-white text-gray-700 placeholder-gray-400 shadow-md rounded-l text-sm;
+  @apply focus:outline-none focus:ring-1 focus:ring-blue-500 focus:border-transparent;
+}
+
+.searchButton {
+  @apply bg-white py-2 px-4 rounded-r;
+  @apply hover:bg-gray-100;
+}
